Fix apply patches

+++ django_mongodb_cli/utils.py
@@ -37,28 +37,29 @@ def apply_patches(app_type):
     """Apply a patch file to the specified project directory."""
     project_dir = project_dirs_map[app_type]
     patch_dir = os.path.join("patches", app_type)
-    for patch_file in os.listdir(patch_dir):
-        shutil.copyfile(
-            os.path.join(patch_dir, patch_file),
-            os.path.join(project_dir, patch_file),
-        )
-        click.echo(click.style(f"Applying patch {patch_file}", fg="blue"))
-        # Ensure the repository is valid
-        repo = Repo(project_dir)
-        if not repo.bare:
-            try:
-                # Apply the patch
-                repo.git.apply(patch_file)
-                click.echo(
-                    f"Patch {os.path.basename(patch_file)} applied successfully."
-                )
-            except Exception as e:
-                click.echo(f"Failed to apply patch: {e}")
+    if os.path.exists(patch_dir):
+        for patch_file in os.listdir(patch_dir):
+            shutil.copyfile(
+                os.path.join(patch_dir, patch_file),
+                os.path.join(project_dir, patch_file),
+            )
+            click.echo(click.style(f"Applying patch {patch_file}", fg="blue"))
+            # Ensure the repository is valid
+            repo = Repo(project_dir)
+            if not repo.bare:
+                try:
+                    # Apply the patch
+                    repo.git.apply(patch_file)
+                    click.echo(
+                        f"Patch {os.path.basename(patch_file)} applied successfully."
+                    )
+                except Exception as e:
+                    click.echo(f"Failed to apply patch: {e}")
+                    return
+            else:
+                click.echo("Not a valid Git repository.")
                 return
-        else:
-            click.echo("Not a valid Git repository.")
-            return
-        click.echo(click.style("Patch applied", fg="green"))
+            click.echo(click.style("Patch applied", fg="green"))
 
 
 def copy_mongo_apps(test_dir, app_type):
